feat: cleanup work on plugins system
