--- src/plugins/hello.ts
@@ -1,3 +1,0 @@
-export async function init() {
-  console.log('Hello, world!')
-}
