+++ src/plugins/themes.ts
@@ -1,0 +1,5 @@
+import { enableCurrentTheme } from "@/seqta/ui/themes/enableCurrent";
+
+export async function init() {
+  enableCurrentTheme();
+}
